Moved payload under the model folder
