+++ tests/UKHO.ERPFacade.API.FunctionalTests/Model/JsonPayloadHelper.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UKHO.ERPFacade.API.FunctionalTests.Helpers
+namespace UKHO.ERPFacade.API.FunctionalTests.Model
 {
     public class JsonPayloadHelper
     {
